+++ test/utils/test-utils.js
@@ -18,6 +18,24 @@ exports.consoleError = function(block, regex) {
     expect(messages).toMatch(regex)
 }
 
+exports.consoleWarn = function(block, regex) {
+    let messages = ""
+    const orig = console.warn
+    console.warn = function() {
+        Object.keys(arguments).forEach(key => {
+            messages += ", " + arguments[key]
+        })
+        messages += "\n"
+    }
+    try {
+        block()
+    } finally {
+        console.warn = orig
+    }
+    expect(messages.length).toBeGreaterThan(0)
+    expect(messages).toMatch(regex)
+}
+
 exports.supressConsole = function(block) {
     const { warn, error } = console
     Object.assign(console, {